Move all problem related classes to exception package
