+++ winter/web/exceptions/problem_annotation.py
@@ -1,0 +1,11 @@
+from http import HTTPStatus
+
+from dataclasses import dataclass
+
+
+@dataclass(frozen=True)
+class ProblemAnnotation:
+    status: HTTPStatus
+    title: str
+    detail: str
+    type: str
